+++ tests/testthat/test_operations_setcomplement.R
@@ -61,6 +61,16 @@ test_that("interval", {
     "[1,10] \\ {2,...,4}"
   )
   expect_equal(Interval$new(1, 5) - Set$new(3, 5), Interval$new(1, 3, type = "[)") + Interval$new(3, 5, type = "()"))
+
+  a = Interval$new(0, 2, type = "[]")
+  expect_equal(a - Interval$new(0, 1, type = "[]"), Interval$new(1, 2, type = "(]"))
+  expect_equal(a - Interval$new(0, 1, type = "[)"), Interval$new(1, 2, type = "[]"))
+  expect_equal(a - Interval$new(0, 1, type = "()"), Set$new(0) + Interval$new(1, 2, type = "[]"))
+  expect_equal(a - Interval$new(0, 1, type = "(]"), Set$new(0) + Interval$new(1, 2, type = "(]"))
+
+  a = Interval$new(1, 3, type = "[]")
+  expect_equal(a - Interval$new(0, 2, type = "()"), Interval$new(2, 3, type = "[]"))
+  expect_equal(a - Interval$new(0, 2, type = "[]"), Interval$new(2, 3, type = "(]"))
 })
 
 test_that("fuzzy", {